+++ plib/hooks/CustomButtons.php
@@ -6,43 +6,44 @@ class Modules_SupervisorManager_CustomButtons extends pm_Hook_CustomButtons
     {
         $icon = pm_Context::getBaseUrl() . 'images/icon.svg';
         $link = pm_Context::getBaseUrl() . 'index.php/index/index';
+        $domainLink = pm_Context::getBaseUrl() . 'index.php/index/domain';
 
-        $buttons = array(
-            array(
-                'place' => self::PLACE_ADMIN_NAVIGATION,
-                'section' => self::SECTION_NAV_SERVER_MANAGEMENT,
-                'title' => 'Supervisor',
-                'description' => 'Manage Supervisor programs',
-                'icon' => $icon,
-                'link' => $link,
-                'newWindow' => false,
-            ),
-            array(
-                'place' => self::PLACE_RESELLER_NAVIGATION,
-                'section' => self::SECTION_NAV_ADDITIONAL,
-                'title' => 'Supervisor',
-                'description' => 'Manage assigned Supervisor programs',
-                'icon' => $icon,
-                'link' => $link,
-                'newWindow' => false,
-            ),
-            array(
-                'place' => self::PLACE_CUSTOMER_HOME,
-                'title' => 'Supervisor',
-                'description' => 'Manage assigned Supervisor programs',
-                'icon' => $icon,
-                'link' => $link,
-                'newWindow' => false,
-            ),
-            array(
-                'place' => self::PLACE_DOMAIN,
-                'title' => 'Supervisor',
-                'description' => 'Manage domain Supervisor programs',
-                'icon' => $icon,
-                'link' => $link,
-                'newWindow' => false,
-                'contextParams' => true,
-            ),
+        $buttons = array();
+        $buttons[] = array(
+            'place' => self::PLACE_ADMIN_NAVIGATION,
+            'section' => self::SECTION_NAV_SERVER_MANAGEMENT,
+            'title' => 'Supervisor',
+            'description' => 'Manage Supervisor programs',
+            'icon' => $icon,
+            'link' => $link,
+            'newWindow' => false,
+        );
+
+        $buttons[] = array(
+            'place' => self::PLACE_RESELLER_NAVIGATION,
+            'section' => self::SECTION_NAV_ADDITIONAL,
+            'title' => 'Supervisor',
+            'description' => 'Manage assigned Supervisor programs',
+            'icon' => $icon,
+            'link' => $link,
+            'newWindow' => false,
+        );
+        $buttons[] = array(
+            'place' => self::PLACE_CUSTOMER_HOME,
+            'title' => 'Supervisor',
+            'description' => 'Manage assigned Supervisor programs',
+            'icon' => $icon,
+            'link' => $link,
+            'newWindow' => false,
+        );
+        $buttons[] = array(
+            'place' => self::PLACE_DOMAIN,
+            'title' => 'Supervisor',
+            'description' => 'Manage domain Supervisor programs',
+            'icon' => $icon,
+            'link' => $domainLink,
+            'newWindow' => false,
+            'contextParams' => true,
         );
 
         if (
@@ -55,7 +56,7 @@ class Modules_SupervisorManager_CustomButtons extends pm_Hook_CustomButtons
                 'title' => 'Supervisor',
                 'description' => 'Manage Supervisor programs',
                 'icon' => $icon,
-                'link' => $link,
+                'link' => $domainLink,
                 'newWindow' => false,
                 'contextParams' => true,
             );